fix: Fix FQDN hostname on FreeBSD
